fix: change computing integration in the app
