+++ src/app/Routes.tsx
@@ -3,9 +3,7 @@ import { BrowserRouter, Redirect, Switch } from 'react-router-dom';
 import HomeView from 'views/HomeView';
 import LoginView from 'views/LoginView';
 import SignupView from 'views/SignupView';
-import RedirectionView from 'views/RedirectionView';
 import DashboardView from 'views/DashboardView';
-import ComputingView from 'views/ComputingView';
 
 import AuthRoute from './AuthRoute';
 import PrivateRoute from './PrivateRoute';
@@ -14,8 +12,6 @@ const Routes = (): JSX.Element => (
 	<BrowserRouter>
 		<Switch>
 			<PrivateRoute path="/dashboard" children={<DashboardView />} />
-			<PrivateRoute path="/computing" children={<ComputingView />} />
-			<PrivateRoute path="/redirection" children={<RedirectionView />} />
 			<AuthRoute exact path="/" children={<HomeView />} />
 			<AuthRoute path="/signup" children={<SignupView />} />
 			<AuthRoute path="/login" children={<LoginView />} />
