+++ src/views/DashboardView.tsx
@@ -19,7 +19,7 @@ import EthCrypto from 'eth-crypto';
 
 import { useUserContext } from 'contexts/user';
 
-import { IPCFile, IPCContact } from 'types/types';
+import { IPCFile, IPCContact, IPCProgram } from 'types/types';
 
 import Modal from 'components/Modal';
 
@@ -37,6 +37,8 @@ const Dashboard = (): JSX.Element => {
 	const { isOpen: isOpenContactAdd, onOpen: onOpenContactAdd, onClose: onCloseContactAdd } = useDisclosure();
 	const { isOpen: isOpenContactUpdate, onOpen: onOpenContactUpdate, onClose: onCloseContactUpdate } = useDisclosure();
 	const { isOpen: isOpenShare, onOpen: onOpenShare, onClose: onCloseShare } = useDisclosure();
+	const { isOpen: isOpenProgram, onOpen: onOpenProgram, onClose: onCloseProgram } = useDisclosure();
+	const [programs, setPrograms] = useState<IPCProgram[]>([]);
 	const [files, setFiles] = useState<IPCFile[]>([]);
 	const [sharedFiles, setSharedFiles] = useState<IPCFile[]>([]);
 	const [contacts, setContacts] = useState<IPCContact[]>([]);
@@ -49,6 +51,7 @@ const Dashboard = (): JSX.Element => {
 	const [selectedTab, setSelectedTab] = useState(0);
 	const [isUploadLoading, setIsUploadLoading] = useState(false);
 	const [isDownloadLoading, setIsDownloadLoading] = useState(false);
+	const [isDeployLoading, setIsDeployLoading] = useState(false);
 	const [fileEvent, setFileEvent] = useState<ChangeEvent<HTMLInputElement> | undefined>(undefined);
 	const [contactsNameEvent, setContactNameEvent] = useState<ChangeEvent<HTMLInputElement> | undefined>(undefined);
 	const [contactsPublicKeyEvent, setContactPublicKeyEvent] = useState<ChangeEvent<HTMLInputElement> | undefined>(
@@ -88,6 +91,54 @@ const Dashboard = (): JSX.Element => {
 				isClosable: true,
 			});
 		}
+	};
+
+	const uploadProgram = async () => {
+		if (!fileEvent || !fileEvent.target.files) return;
+		const filename = extractFilename(fileEvent.target.value);
+
+		if (!filename) return;
+
+		setIsDeployLoading(true);
+		try {
+			if (user.account) {
+				const upload = await user.computing.uploadProgram(
+					{
+						name: filename,
+						hash: '',
+						created_at: Date.now(),
+					},
+					fileEvent.target.files[0],
+				);
+				toast({
+					title: upload.message,
+					status: upload.success ? 'success' : 'error',
+					duration: 2000,
+					isClosable: true,
+				});
+				if (upload.success) {
+					setPrograms(user.computing.programs);
+				}
+			} else {
+				toast({
+					title: 'Failed to load account',
+					status: 'error',
+					duration: 2000,
+					isClosable: true,
+				});
+			}
+			onClose();
+		} catch (error) {
+			console.error(error);
+			toast({
+				title: 'Unable to upload file',
+				status: 'error',
+				duration: 2000,
+				isClosable: true,
+			});
+		}
+		setFileEvent(undefined);
+		setIsDeployLoading(false);
 	};
 
 	const uploadFile = async () => {
@@ -147,6 +198,7 @@ const Dashboard = (): JSX.Element => {
 				isClosable: true,
 			});
 		}
+		setFileEvent(undefined);
 		setIsUploadLoading(false);
 	};
 
@@ -326,14 +378,17 @@ const Dashboard = (): JSX.Element => {
 		<HStack minH="100vh" minW="100vw" align="start">
 			<ResponsiveBar
 				onOpen={onOpen}
+				onOpenProgram={onOpenProgram}
 				setSelectedTab={setSelectedTab}
 				isUploadLoading={isUploadLoading}
+				isDeployLoading={isDeployLoading}
 				selectedTab={selectedTab}
 			/>
 			<Box w="100%" m="32px !important">
 				<VStack w="100%" maxW="350px" id="test" spacing="16px" mt={{ base: '64px', lg: '0px' }}>
 					<DisplayFileCards
 						myFiles={files}
+						myPrograms={programs}
 						sharedFiles={sharedFiles}
 						contacts={contacts}
 						index={selectedTab}
@@ -348,6 +403,32 @@ const Dashboard = (): JSX.Element => {
 					/>
 				</VStack>
 			</Box>
+			<Modal
+				isOpen={isOpenProgram}
+				onClose={onCloseProgram}
+				title="Deploy a program"
+				CTA={
+					<Button
+						variant="inline"
+						w="100%"
+						mb="16px"
+						onClick={uploadProgram}
+						isLoading={isDeployLoading}
+						id="ipc-dashboardView-upload-program-modal-button"
+					>
+						Deploy program
+					</Button>
+				}
+			>
+				<Input
+					type="file"
+					h="100%"
+					w="100%"
+					p="10px"
+					onChange={(e: ChangeEvent<HTMLInputElement>) => setFileEvent(e)}
+					id="ipc-dashboardView-upload-program"
+				/>
+			</Modal>
 			<Modal
 				isOpen={isOpen}
 				onClose={onClose}
